+++ CHANGELOG.md
@@ -15,6 +15,9 @@ All notable changes to this project will be documented in this file.
 - Added support for `better-sqlite3` without knex
 - Added support for `bun:sqlite`
 - Added a generic implementation for `sqlite` that can be easily extended to any provider
+- Added `ForbiddenError` to the api
+- Added `apiPreprocessFilter` and `backendPreprocessFilter
+- Added a way to analyze filter and query it - `Filter.getInfo`
 - Fixed an issue with entity ids that included date
 - Fixed an issue with chaining of validators that in some cases caused a validator to be overwritten
 - Changed postgres schema builder to use `timestamptz` instead of `timestamp`